# .migration-backup/lib/api-zod/src/generated/types/jobStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Job dispatch API for transport booking and driver claims
 * OpenAPI spec version: 0.1.0
 */

export interface JobStats {
  total: number;
  pending: number;
  pendingDeposit: number;
  depositReceived: number;
  claimed: number;
  completed: number;
  expired: number;
  cancelled: number;
}